+++ tests/TestCase.php
@@ -2,7 +2,12 @@
 
 namespace Nesk\Puphpeteer\Tests;
 
+use Monolog\Logger;
+use ReflectionClass;
+use Psr\Log\LogLevel;
+use PHPUnit\Framework\Constraint\Callback;
 use PHPUnit\Framework\TestCase as BaseTestCase;
+use PHPUnit\Framework\MockObject\Matcher\Invocation;
 
 class TestCase extends BaseTestCase
 {
@@ -23,5 +28,43 @@ class TestCase extends BaseTestCase
     public function canPopulateProperty(string $propertyName): bool
     {
         return !in_array($propertyName, $this->dontPopulateProperties);
+    }
+
+    public function loggerMock($expectations) {
+        $loggerMock = $this->getMockBuilder(Logger::class)
+            ->setConstructorArgs(['rialto'])
+            ->setMethods(['log'])
+            ->getMock();
+
+        if ($expectations instanceof Invocation) {
+            $expectations = [func_get_args()];
+        }
+
+        foreach ($expectations as $expectation) {
+            [$matcher] = $expectation;
+            $with = array_slice($expectation, 1);
+
+            $loggerMock->expects($matcher)
+                ->method('log')
+                ->with(...$with);
+        }
+
+        return $loggerMock;
+    }
+
+    public function isLogLevel(): Callback {
+        $psrLogLevels = (new ReflectionClass(LogLevel::class))->getConstants();
+        $monologLevels = (new ReflectionClass(Logger::class))->getConstants();
+        $monologLevels = array_intersect_key($monologLevels, $psrLogLevels);
+
+        return $this->callback(function ($level) use ($psrLogLevels, $monologLevels) {
+            if (is_string($level)) {
+                return in_array($level, $psrLogLevels, true);
+            } else if (is_int($level)) {
+                return in_array($level, $monologLevels, true);
+            }
+
+            return false;
+        });
     }
 }